// ui/src/generated/InstrumentTypeInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InstrumentType } from "./InstrumentType";
import type { KnobPage } from "./KnobPage";
import type { OutputInfo } from "./OutputInfo";
import type { ParamInfo } from "./ParamInfo";

export type InstrumentTypeInfo = { type: InstrumentType, label: string, default_id: string, 
/**
 * Outputs of an instance with the default id.
 */
outputs: Array<OutputInfo>, 
/**
 * Parameters of an instance with the default id.
 */
params: Array<ParamInfo>, 
/**
 * Knob pages of an instance with the default id.
 */
knob_pages: Array<KnobPage>, };